Clean Repository
Move all files to the paper bin folder to make space for a easier
handling of files
