+++ PaperBin/Assumptions Worksheet_161017_v1.docx
