+++ PaperBin/ColorGrid.py
